Fixed execution of Node Update.
if it hasn't started yet and it is in the running state run OnStart and return the state of the node.

if the node is running (The State of the node is not Failure or success then it is still running) run OnUpdate and retun that the Node is Running.

If we have gotten here then the node is not running.
If the Node has Stopped (started is false) return the Nodes state.

If we have gotten here the Node is not Running and Has Not Stopped. Set started to false run On Stop. return node is running.